refactor(api): move unwrapDataEnvelope to schema utils
