+++ api/src/controllers/schema.ts
@@ -13,34 +13,12 @@ import { SchemaService } from '../services/schema.js';
 import asyncHandler from '../utils/async-handler.js';
 import { getVersionedHash } from '../utils/get-versioned-hash.js';
 import { queryFlag } from '../utils/query-flag.js';
+import { unwrapDataEnvelope } from '../utils/schema/unwrap-data-envelope.js';
 
 const env = useEnv();
 const router = express.Router();
 
 const IMPORT_MAX_FILE_SIZE = bytes.parse(env['IMPORT_MAX_FILE_SIZE'] as string) ?? undefined;
-
-/**
- * Unwrap the standard Directus API response envelope (`{ data: T }`) from a request body.
- *
- * When a user pipes the JSON output of `GET /schema/snapshot` directly into
- * `POST /schema/diff` or `POST /schema/apply`, the body is already wrapped in
- * `{ "data": <payload> }` by the respond middleware. Both the raw form and the
- * envelope form are accepted transparently.
- */
-export function unwrapDataEnvelope<T extends object>(body: T | { data: T }): T {
-	if (
-		body !== null &&
-		typeof body === 'object' &&
-		'data' in body &&
-		Object.keys(body).length === 1 &&
-		typeof (body as { data?: unknown }).data === 'object' &&
-		(body as { data?: unknown }).data !== null
-	) {
-		return (body as { data: T }).data;
-	}
-
-	return body as T;
-}
 
 /** Accepts a single collection name or a list, always normalizing to a string array. */
 const collectionList = z.union([z.string(), z.array(z.string())]).transform((value) => toArray(value));
